chore: sd-headline docs
